Create RDF and XML metadata for new sensor manufacturers

Wrapper function that automates the generation of both RDF/Turtle (.ttl) and SensorML/XML (.xml) metadata files for new sensor manufacturers. It takes as input a manufacturer table (man_table) and the ORCID identifier of the metadata creator, and then calls the internal functions sensors_create_rdf_ttl_manufacturer() and sensors_create_rdf_XML_manufacturer() in sequence.

Usage

sensors_new_manufacturer(man_table = NULL, orcid_creator = NULL)

Arguments

man_table

A data frame containing manufacturer metadata (name, contact, address, etc.).

orcid_creator

Character string with the ORCID URI of the metadata creator.

Value

Creates and saves RDF/Turtle and XML files describing each manufacturer. No object is returned.

Author

Alessandro Oggioni, phD (2023) oggioni.a@cnr.it

Examples

# Example of a manufacturer entry
manufacturer_list <- tibble::tibble(
  name = "New Manufacturer",
  homepage = "https://www.newmanufacturer.com",
  email = "info@newmanufacturer.com",
  phone = "+39 02 1234567",
  thoroughfareName = "Via Roma",
  streetNr = "10",
  postcode = "20100",
  building = NA,
  town = "Milano",
  region = "Lombardia",
  country = "Italy"
)
sensors_new_manufacturer(
  man_table = manufacturer_list,
  orcid_creator = "0000-0002-1234-5678"
)
#> Error in sensors_new_manufacturer(man_table = manufacturer_list, orcid_creator = "0000-0002-1234-5678"): could not find function "sensors_new_manufacturer"